+++ IntegraJeff/IntegraJeff/FrmIntermediario.cs
@@ -33,9 +33,9 @@ namespace IntegraJeff
         {
             InitializeComponent();
             IntermediarioActual = enviado;
-            btnAgregar.Enabled = false;
-            btnEditar.Enabled = false;
-            btnEliminar.Enabled = false;
+            //btnAgregar.Enabled = false;
+            //btnEditar.Enabled = false;
+            //btnEliminar.Enabled = false;
         }
 
         private void FrmIntermediario_Load(object sender, EventArgs e)
@@ -49,26 +49,42 @@ namespace IntegraJeff
 
         private void cboPagina_SelectionChangeCommitted(object sender, EventArgs e)
         {
-            varPagIndice = cboPagina.SelectedIndex;
-            if (chkBusqueda.Checked)
+            try
             {
-                ActualizarPaginasBusqueda();
+                varPagIndice = cboPagina.SelectedIndex;
+                if (chkBusqueda.Checked)
+                {
+                    ActualizarPaginasBusqueda();
+                }
+                else
+                {
+                    ActualizarPaginasNormal();
+                }
             }
-            else
+            catch (Exception ex)
             {
-                ActualizarPaginasNormal();
+
+                MessageBox.Show(ex.Message);
             }
         }
 
         private void btnRefresh_Click(object sender, EventArgs e)
         {
-            if (chkBusqueda.Checked)
+            try
             {
-                ActualizarPaginasBusqueda();
+                if (chkBusqueda.Checked)
+                {
+                    ActualizarPaginasBusqueda();
+                }
+                else
+                {
+                    ActualizarPaginasNormal();
+                }
             }
-            else
+            catch (Exception ex)
             {
-                ActualizarPaginasNormal();
+
+                MessageBox.Show(ex.Message);
             }
         }
 
@@ -80,15 +96,23 @@ namespace IntegraJeff
 
         private void chkBusqueda_CheckedChanged(object sender, EventArgs e)
         {
-            cboPagina.SelectedIndex = 0;
-            varPagIndice = cboPagina.SelectedIndex;
-            if (chkBusqueda.Checked)
+            try
             {
-                ActualizarPaginasBusqueda();
+                cboPagina.SelectedIndex = 0;
+                varPagIndice = cboPagina.SelectedIndex;
+                if (chkBusqueda.Checked)
+                {
+                    ActualizarPaginasBusqueda();
+                }
+                else
+                {
+                    ActualizarPaginasNormal();
+                }
             }
-            else
+            catch (Exception ex)
             {
-                ActualizarPaginasNormal();
+
+                MessageBox.Show(ex.Message);
             }
         }
 
@@ -106,30 +130,37 @@ namespace IntegraJeff
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("¿Desea eliminar este cliente?", "ELIMINAR", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            if (ConexionBD.UsuarioActual.Rol.Equals("Administrador"))
             {
-                MySqlConnection conn = ConexionBD.Conexion;
-                try
+                if (MessageBox.Show("¿Desea eliminar este cliente?", "ELIMINAR", MessageBoxButtons.YesNo) == DialogResult.Yes)
                 {
-                    string strOrder = "DELETE FROM `integrajeffschm`.`intermediarios` WHERE `idIntermediarios` = @id;";
-                    MySqlCommand mscCommand = new MySqlCommand(strOrder, conn);
-                    mscCommand.Parameters.AddWithValue("@id", IntermediarioActual.IDIntermediario);
-                    conn.Open();
-                    mscCommand.ExecuteNonQuery();
-                }
-                catch (Exception ex)
-                {
+                    MySqlConnection conn = ConexionBD.Conexion;
+                    try
+                    {
+                        string strOrder = "DELETE FROM `integrajeffschm`.`intermediarios` WHERE `idIntermediarios` = @id;";
+                        MySqlCommand mscCommand = new MySqlCommand(strOrder, conn);
+                        mscCommand.Parameters.AddWithValue("@id", IntermediarioActual.IDIntermediario);
+                        conn.Open();
+                        mscCommand.ExecuteNonQuery();
+                    }
+                    catch (Exception ex)
+                    {
 
-                    MessageBox.Show(ex.Message);
-                }
-                finally
-                {
-                    conn.Close();
-                }
-                IntermediarioActual.IDIntermediario = 0;
-                IntermediarioActual.Nombre = "";
-                txtIdIntermediario.Text = IntermediarioActual.IDIntermediario.ToString();
-                txtNombre.Text = IntermediarioActual.Nombre;
+                        MessageBox.Show(ex.Message);
+                    }
+                    finally
+                    {
+                        conn.Close();
+                    }
+                    IntermediarioActual.IDIntermediario = 0;
+                    IntermediarioActual.Nombre = "";
+                    txtIdIntermediario.Text = IntermediarioActual.IDIntermediario.ToString();
+                    txtNombre.Text = IntermediarioActual.Nombre;
+                } 
+            }
+            else
+            {
+                MessageBox.Show("Usted no puede eliminar porque no es administrador");
             }
         }
         void ActualizarPaginasNormal()
